Backend/pyproject.toml
[tool.poetry]
name = "rag-backend"
version = "0.1.0"
description = "Production-grade RAG backend for tabular data"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = "^0.29"
pydantic = "^2.6"
pydantic-settings = "^2.2"
structlog = "^24.1"
opentelemetry-api = "^1.23"
opentelemetry-sdk = "^1.23"
qdrant-client = "^1.8"
python-multipart = "^0.0.21"
sqlalchemy = "^2.0.45"
alembic = "^1.17.2"
psycopg = "^3.3.2"
pandas = "^2.3.3"
openpyxl = "^3.1.5"
python-dotenv = "^1.2.1"
ollama = "^0.6.1"
python-docx = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1"
ruff = "^0.3"
mypy = "^1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
